Sync upstream

+++ presto-raptor/src/main/java/com/facebook/presto/raptor/storage/Row.java
@@ -42,9 +42,9 @@ import static java.util.Objects.requireNonNull;
 public class Row
 {
     private final List<Object> columns;
-    private final int sizeInBytes;
+    private final long sizeInBytes;
 
-    public Row(List<Object> columns, int sizeInBytes)
+    public Row(List<Object> columns, long sizeInBytes)
     {
         this.columns = requireNonNull(columns, "columns is null");
         checkArgument(sizeInBytes >= 0, "sizeInBytes must be >= 0");
@@ -56,7 +56,7 @@ public class Row
         return columns;
     }
 
-    public int getSizeInBytes()
+    public long getSizeInBytes()
     {
         return sizeInBytes;
     }
@@ -70,7 +70,7 @@ public class Row
         for (int channel = 0; channel < page.getChannelCount(); channel++) {
             Block block = page.getBlock(channel);
             Type type = types.get(channel);
-            int size;
+            long size;
             Object value = getNativeContainerValue(type, block, position);
             if (value == null) {
                 size = SIZE_OF_BYTE;
@@ -180,7 +180,7 @@ public class Row
 
     private static class RowBuilder
     {
-        private int rowSize;
+        private long rowSize;
         private final List<Object> columns;
 
         public RowBuilder(int columnCount)
@@ -188,7 +188,7 @@ public class Row
             this.columns = new ArrayList<>(columnCount);
         }
 
-        public void add(Object value, int size)
+        public void add(Object value, long size)
         {
             columns.add(value);
             rowSize += size;
